:heavy_check_mark: 4.45 Updating exercise.

+++ Chapter 4/4.45/bubble_sort_pointer.c
@@ -19,7 +19,7 @@ void bubble_s_pointer(long *data, long count) {
 
 int main()
 {
-    bubble_s_pointer(array, 10);
+    bubble_s_pointer(array, (long)10);
 
     return 0;
 }
